fixes #926 and #928
