+++ src/index.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 import {FocusStyleManager} from "@blueprintjs/core";
-import {AppContainer} from "./App";
+import {App} from "./App";
 import {unregister} from "./registerServiceWorker";
 import "./index.css";
 import "@blueprintjs/core/lib/css/blueprint.css";
@@ -29,7 +29,7 @@ window["React"] = React; // tslint:disable-line
 window["ReactDOM"] = ReactDOM; // tslint:disable-line
 
 ReactDOM.render(
-    <AppContainer/>,
+    <App/>,
     document.getElementById("root") as HTMLElement
 );
